Fix Page numbering for combined statements PDF
