+++ CmsWeb/CMSWeb.csproj
@@ -546,7 +546,10 @@
     <Compile Include="Areas\Finance\Models\DepositsModel.cs" />
     <Compile Include="Areas\Finance\Models\DepositTotalsModel.cs" />
     <Compile Include="Areas\Finance\Models\StatementSpecification.cs" />
+    <Compile Include="Areas\Finance\Models\Statements\FooterOptions.cs" />
+    <Compile Include="Areas\Finance\Models\Statements\HeaderOptions.cs" />
     <Compile Include="Areas\Finance\Models\Statements\ListOfNormalContributions.cs" />
+    <Compile Include="Areas\Finance\Models\Statements\MarginOptions.cs" />
     <Compile Include="Areas\Finance\Models\Statements\StatementContext.cs" />
     <Compile Include="Areas\Finance\Models\Statements\StatementOptions.cs" />
     <Compile Include="Areas\Finance\Models\RemoteDepositCapture.cs" />
